Front-End/frontend/Server/auth_server/requirements.txt
bleach==3.1.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
docker==4.2.1
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
requests==2.23.0
six==1.15.0
urllib3==1.25.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1